Iris-Flower-Analysis-AI
In this comprehensive project, we delve into the fascinating realm of Artificial Intelligence, specifically focusing on the analysis of the Iris flower using machine learning and deep learning techniques. The dataset chosen for this exploration is the renowned "Iris" dataset, celebrated for its suitability in the field.

Our primary objective is to employ advanced machine learning methodologies to discern and classify the three distinct Iris plant species: Iris setosa, Iris virginica, and Iris versicolor. Leveraging features such as sepal length, sepal width, petal length, and petal width, we aim to train models capable of accurately predicting the species of an Iris flower based on its measured attributes.

Additionally, we will explore the potential of deep learning techniques, seeking to enhance our analysis through neural networks. The project unfolds with data preprocessing, feature scaling, and model training, followed by an in-depth examination of model performance and fine-tuning for optimal results.